refactor(udp-tracker-server): [#1589] move average processing time calculation from Repository to Metrics
