+++ packages/udp-tracker-server/src/statistics/repository.rs
@@ -73,84 +73,32 @@ impl Repository {
         result
     }
 
-    #[allow(clippy::cast_precision_loss)]
-    #[allow(clippy::cast_possible_truncation)]
-    #[allow(clippy::cast_sign_loss)]
     pub async fn recalculate_udp_avg_connect_processing_time_ns(&self, req_processing_time: Duration) -> f64 {
         let stats_lock = self.stats.write().await;
 
-        let req_processing_time = req_processing_time.as_nanos() as f64;
-        let udp_connections_handled = (stats_lock.udp4_connections_handled() + stats_lock.udp6_connections_handled()) as f64;
-
-        let previous_avg = stats_lock.udp_avg_connect_processing_time_ns();
-
-        // Moving average: https://en.wikipedia.org/wiki/Moving_average
-        let new_avg = previous_avg as f64 + (req_processing_time - previous_avg as f64) / udp_connections_handled;
+        let new_avg = stats_lock.recalculate_udp_avg_connect_processing_time_ns(req_processing_time);
 
         drop(stats_lock);
-
-        tracing::debug!(
-            "Recalculated UDP average connect processing time: {} ns (previous: {} ns, req_processing_time: {} ns, udp_connections_handled: {})",
-            new_avg,
-            previous_avg,
-            req_processing_time,
-            udp_connections_handled
-        );
 
         new_avg
     }
 
-    #[allow(clippy::cast_precision_loss)]
-    #[allow(clippy::cast_possible_truncation)]
-    #[allow(clippy::cast_sign_loss)]
     pub async fn recalculate_udp_avg_announce_processing_time_ns(&self, req_processing_time: Duration) -> f64 {
         let stats_lock = self.stats.write().await;
 
-        let req_processing_time = req_processing_time.as_nanos() as f64;
-
-        let udp_announces_handled = (stats_lock.udp4_announces_handled() + stats_lock.udp6_announces_handled()) as f64;
-
-        let previous_avg = stats_lock.udp_avg_announce_processing_time_ns();
-
-        // Moving average: https://en.wikipedia.org/wiki/Moving_average
-        let new_avg = previous_avg as f64 + (req_processing_time - previous_avg as f64) / udp_announces_handled;
+        let new_avg = stats_lock.recalculate_udp_avg_announce_processing_time_ns(req_processing_time);
 
         drop(stats_lock);
-
-        tracing::debug!(
-            "Recalculated UDP average announce processing time: {} ns (previous: {} ns, req_processing_time: {} ns, udp_announces_handled: {})",
-            new_avg,
-            previous_avg,
-            req_processing_time,
-            udp_announces_handled
-        );
 
         new_avg
     }
 
-    #[allow(clippy::cast_precision_loss)]
-    #[allow(clippy::cast_possible_truncation)]
-    #[allow(clippy::cast_sign_loss)]
     pub async fn recalculate_udp_avg_scrape_processing_time_ns(&self, req_processing_time: Duration) -> f64 {
         let stats_lock = self.stats.write().await;
 
-        let req_processing_time = req_processing_time.as_nanos() as f64;
-        let udp_scrapes_handled = (stats_lock.udp4_scrapes_handled() + stats_lock.udp6_scrapes_handled()) as f64;
-
-        let previous_avg = stats_lock.udp_avg_scrape_processing_time_ns();
-
-        // Moving average: https://en.wikipedia.org/wiki/Moving_average
-        let new_avg = previous_avg as f64 + (req_processing_time - previous_avg as f64) / udp_scrapes_handled;
+        let new_avg = stats_lock.recalculate_udp_avg_scrape_processing_time_ns(req_processing_time);
 
         drop(stats_lock);
-
-        tracing::debug!(
-            "Recalculated UDP average scrape processing time: {} ns (previous: {} ns, req_processing_time: {} ns, udp_scrapes_handled: {})",
-            new_avg,
-            previous_avg,
-            req_processing_time,
-            udp_scrapes_handled
-        );
 
         new_avg
     }
